pyproject.toml
[project]
name = "paperpulse"
version = "0.1.0"
description = "A daily medical literature digest for your dashboard"
authors = [
    {name = "Engin Boz"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = ["requests (>=2.33.1,<3.0.0)", "python-dotenv (>=1.2.2,<2.0.0)", "sqlalchemy (>=2.0.49,<3.0.0)", "psycopg2-binary (>=2.9.12,<3.0.0)", "flask (>=3.1.3,<4.0.0)", "sentence-transformers (>=5.4.1,<6.0.0)"]

[project.urls]
Repository = "https://github.com/enginboz/paperpulse"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = ["pytest-mock (>=3.15.1,<4.0.0)"]